sepperated and trig methods and separated them into separtate files
