--- test/test_one_arg_calcs.py
@@ -1,108 +1,0 @@
-from unittest import TestCase
-from test.error_colors import ErrorColors
-from calculator import Calculator
-
-
-class TestCalcOneArg(TestCase):
-
-    def _test(self, method, tests):
-        for num, expected in tests:
-            actual = method(num)
-            error_message = self.__return_calc_error(num, expected, actual)
-            self.assertEqual(expected, actual, error_message)
-
-    def __return_calc_error(self, n1, exp, act):
-        result = f"""
-        {ErrorColors.GREEN}Expected: ({n1}) -> {exp}
-        {ErrorColors.RED}Actual:   ({n1}) -> {act}{ErrorColors.RESET}
-        """
-        return result
-
-    # ONE ARG TESTS
-    def test_square(self):
-        c = Calculator()
-        self._test(c.square, [
-            (2, 4), (3, 9), (-3, 9), (5, 25), (10, 100), (8, 64), (7, 49), (-6, 36)
-        ])
-
-    def test_square_root(self):
-        c = Calculator()
-        self._test(c.square_root, [
-            (2, 1.4142135623730951),(4, 2.0),(6, 2.449489742783178),(8, 2.8284271247461903),(10, 3.1622776601683795),
-            (44, 6.6332495807108),(5, 2.23606797749979),(23, 4.795831523312719),(61, 7.810249675906654)
-        ])
-
-    def test_inverse(self):
-        c = Calculator()
-        self._test(c.inverse, [
-            (1, 1.0),(2, 0.5),(3, 0.3333333333333333),(20, 0.05), (30, 0.03333333333333333),
-            (40, 0.025),(-50, -0.02), (-100, -0.01),(-45, -0.022222222222222223)
-        ])
-
-    def test_switch_sign(self):
-        c = Calculator()
-        self._test(c.switch_sign, [
-            (2, -2), (3, -3), (-3, 3), (5, -5), (10, -10), (-8, 8), (-7, 7), (-6, 6)
-        ])
-
-    def test_sine(self):
-        c = Calculator()
-        self._test(c.sine, [
-            (1, 0.8414709848078965),(2, 0.9092974268256817),(3, 0.1411200080598672), 
-            (20, 0.9129452507276277), (30, -0.9880316240928618), (40, 0.7451131604793488), 
-            (-50, 0.26237485370392877), (-100, 0.5063656411097588),(-45, -0.8509035245341184)
-        ])
-
-    def test_cosine(self):
-        c = Calculator()
-        self._test(c.cosine, [
-            (1, 0.5403023058681398),(2, -0.4161468365471424),(3, -0.9899924966004454),
-            (20, 0.40808206181339196),(30, 0.15425144988758405),(40, -0.6669380616522619),
-            (-50, 0.9649660284921133),(-100, 0.862318872287684),(-45, 0.5253219888177297)
-        ])
-
-    def test_tangent(self):
-        c = Calculator()
-        self._test(c.tangent, [
-            (1, 1.557407724654902),(2, -2.185039863261519),(3, -0.1425465430742778),
-            (20, 2.2371609442247427),(30, -6.4053311966462765),(40, -1.117214930923896),
-            (-50, 0.2719006119976307),(-100, 0.587213915156929),(-45, -1.6197751905438615)
-        ])
-
-    def test_inverse_sine(self):
-        c = Calculator()
-        self._test(c.inverse_sine, [
-            (0.55, 0.5823642378687435),(-0.55, -0.5823642378687435),(0, 0.0),
-            (1, 1.5707963267948966),(-1, -1.5707963267948966),(0.22, 0.22181447049679442),
-            (-0.22, -0.22181447049679442),(0.65, 0.7075844367253556),(-0.65, -0.7075844367253556)
-        ])
-
-    def test_inverse_cosine(self):
-        c = Calculator()
-        self._test(c.inverse_cosine, [
-            (0.55, 0.9884320889261531),(-0.55, 2.15316056466364),(0, 1.5707963267948966),(1, 0.0),
-            (-1, 3.141592653589793),(0.22, 1.3489818562981022),(-0.22, 1.792610797291691), 
-            (0.65, 0.863211890069541), (-0.65, 2.278380763520252)
-        ])
-
-    def test_inverse_tangent(self):
-        c = Calculator()
-        self._test(c.inverse_tangent, [
-            (0.55, 0.5028432109278609),(-0.55, -0.5028432109278609),(0, 0.0),(1, 0.7853981633974483),
-            (-1, -0.7853981633974483),(0.22, 0.21655030497608926),(-0.22, -0.21655030497608926),
-            (0.65, 0.5763752205911837),(-0.65, -0.5763752205911837)
-        ])
-
-    def test_factorial(self):
-        c = Calculator()
-        self._test(c.factorial, [
-            (-5, 120),(-1, 1),(1, 1),(2, 2),(0, 1),(6, 720),(4, 24),(9, 362880),(15, 1307674368000)
-        ])
-
-    def test_natural_log(self):
-        c = Calculator()
-        self._test(c.natural_log, [
-            (1, 0.0),(2, 0.6931471805599453),(3, 1.0986122886681098),(20, 2.995732273553991),(35, 3.5553480614894135),
-            (123, 4.812184355372417),(54, 3.9889840465642745),(525, 6.263398262591624),(262, 5.568344503761097),
-            (100, 4.605170185988092)
-        ])
