+++ test/test_trig_units_switch.py
@@ -3,18 +3,18 @@ from test.error_colors import ErrorColors
 from calculator import Calculator
 
 
-class TestDisplaySwitch(TestCase):
+class TestTrigUnitSwitch(TestCase):
 
     def __switch_test(self, calc_object, tests):  # This tests switch display WITH NO arguments
         for expected_mode in tests:
-            current_mode = calc_object.show_current_display_mode()
+            current_mode = calc_object.get_current_trig_unit()
             error_message = self.__return_calc_error(current_mode, expected_mode)
-            calc_object.switch_display_mode()
+            calc_object.switch_units()
             self.assertEqual(current_mode, expected_mode, error_message)
 
     def __set_test(self, calc_object, tests):  # This tests switch display mode WITH arguments
         for input, expected in tests:
-            actual = calc_object.switch_display_mode(input)
+            actual = calc_object.switch_units(input)
             error_message = self.__return_calc_error(actual, expected)
             self.assertEqual(actual, expected, error_message)
     
@@ -26,17 +26,17 @@ class TestDisplaySwitch(TestCase):
         return result
 
     # TWO ARG TESTS
-    def test_display_switch(self):
+    def test_trig_unit_switch(self):
         c = Calculator() # these are the expected ouputs
-        self.__switch_test(c, ['dec', 'hex', 'oct', 'bin', 'dec'])
+        self.__switch_test(c, ['deg', 'rad', 'deg', 'rad', 'deg', 'rad'])
 
-    def test_display_set(self):
+    def test_trig_unit_set(self):
         c = Calculator()
         self.__set_test(c, [
-            ('dec','dec'), ('hex','hex'), ('oct','oct'), ('bin','bin'),
-            (1, 'Err: please enter valid display mode: bin | oct | hex | dec'),
-            ('n', 'Err: please enter valid display mode: bin | oct | hex | dec'),
-            ('ser', 'Err: please enter valid display mode: bin | oct | hex | dec'),
-            ({}, 'Err: please enter valid display mode: bin | oct | hex | dec'),
-            ([], 'Err: please enter valid display mode: bin | oct | hex | dec')
+            ('rad','rad'), ('deg','deg'),
+            (1, 'Err: please enter a valid unit: deg | rad'),
+            ('n', 'Err: please enter a valid unit: deg | rad'),
+            ('ser', 'Err: please enter a valid unit: deg | rad'),
+            ({}, 'Err: please enter a valid unit: deg | rad'),
+            ([], 'Err: please enter a valid unit: deg | rad')
         ])
